view: call handler after fetching mail and set/use data at point
After fetching a mail, the handler is called with data object as
argument.
In our case, the handler is reponsible to insert the mail
in *jmail-view* buffer.

Furthermore the data is now saved in text properties.
All the functions reply, forward, save attachments ... can now read
the data in text properties at point.

Signed-off-by: Julien Masson <massonju.eseo@gmail.com>