Introduce jmail-view-thread
jmail-view-thread is an alternative view used after RET on a email
from jmail search buffer.

If the email is part of an emails thread, all the emails from the
thread are displayed in *jmail-view* buffer.
A new mode jmail-view-thread is activated on the buffer with new
keybindings specific to jmail-view-thread mode.

This feature can be enabled via jmail-view-thread-default-view.

Signed-off-by: Julien Masson <massonju.eseo@gmail.com>